+++ back/src/main/java/org/crunchytorch/coddy/user/exception/handler/impl/ForbiddenMapper.java
@@ -1,17 +1,16 @@
-package org.crunchytorch.coddy.user.exception.mapper;
+package org.crunchytorch.coddy.user.exception.handler.impl;
 
-import javax.ws.rs.NotAuthorizedException;
+import javax.ws.rs.ForbiddenException;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 import javax.ws.rs.ext.ExceptionMapper;
 import javax.ws.rs.ext.Provider;
 
 @Provider
-public class NotAuthorizedMapper implements ExceptionMapper<NotAuthorizedException> {
-
+public class ForbiddenMapper implements ExceptionMapper<ForbiddenException> {
     @Override
-    public Response toResponse(NotAuthorizedException exception) {
-        return Response.status(Response.Status.UNAUTHORIZED)
+    public Response toResponse(ForbiddenException exception) {
+        return Response.status(Response.Status.FORBIDDEN)
                 .entity(new org.crunchytorch.coddy.application.data.Response(exception.getMessage()))
                 .type(MediaType.APPLICATION_JSON)
                 .build();
